Separate tests into app/ and unit/
