+++ package.json
@@ -17,7 +17,7 @@
     "gcp-build": "node build.js",
     "deploy": "gcloud app deploy",
     "start": "node app.js",
-    "test": "NODE_ENV=test nyc mocha"
+    "test": "NODE_ENV=test nyc mocha --recursive test"
   },
   "dependencies": {
     "@octokit/rest": "^16.17.0",
